% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api-key.R
\name{trem_set_api_key}
\alias{trem_set_api_key}
\title{Set Tremendous API Key in .Renviron}
\usage{
trem_set_api_key(api_key)
}
\arguments{
\item{api_key}{API key from
\href{https://developers.tremendous.com/}{tremendous.com}.}
}
\value{
NA; Used to set the Tremendous API key for persistent use in '.Renviron'.
}
\description{
Set Tremendous API Key in .Renviron
}
\examples{

if (interactive()) {
  # For Sandbox Environment, your API key should look like:
  trem_set_api_key("TEST_YOUR-API-KEY")

  # For Production Environment, your API key should look like:
  trem_set_api_key("PROD_YOUR-API-KEY")
}

}
