% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{trem_ua}
\alias{trem_ua}
\title{Get useragent info for tremendousr API Package}
\usage{
trem_ua()
}
\value{
useragent info for tremendousr API Package
}
\description{
Adapted from \href{https://github.com/jdtrat/chimpr/}{chimpr}.
}
